goto command proof of conceptworks

+++ uartParser.h
@@ -1,6 +1,6 @@
 void copy_command(void);
 void process_command(void);
-void print_value (char *id, int value);
+void print_value (char *id, double value);
 void uart_ok(void);
 double parse_assignment (char input[]);
 void process_uart(void);
